Disable state for dry run
Dry run shouldn't need this.
Furthermore it also make a dryrun less complicated.

This uses optional string to get rid of uploading
state where appropriate as discussed here:
https://github.com/hercules-ci/hercules-ci-effects/pull/18#discussion_r604057874

This also works around a rare bug where
curl sometimes gives a 500 error on uploading
the state.
(Which is important to our use case)

--- squash log

Disable the check for creating the deployment

Uncondtionally define stateFileName

This allows postStateScript to use it as well